+++ scengen/generator.py
@@ -9,6 +9,7 @@ from pathlib import Path
 from typing import Union, List, Dict, Any, Tuple
 
 from fameio.source.loader import load_yaml
+from fameio.source.tools import keys_to_lower
 
 from scengen.cli import CreateOptions
 from scengen.files import write_yaml, save_seed_to_trace_file, get_trace_file
@@ -38,6 +39,7 @@ ERR_NO_INTEGER = "Expected a single integer or '{}' but received '{}' for `agent
 ERR_COULD_NOT_MAP_RANGE_VALUES = "Could not map range values '{}' to minimum, maximum values."
 ERR_FAILED_RESOLVE_ID = "Found replacement Identifier '{}' with no corresponding Agent in Contract '{}'"
 DEBUG_NO_CREATE = "No agents to `create` found in Config '{}'"
+DEBUG_NO_PATH_TO_BE_REPLACED_IN = "No path to be replaced for Attribute '{}: {}'."
 
 
 def generate_scenario(options: dict) -> None:
@@ -45,19 +47,16 @@ def generate_scenario(options: dict) -> None:
     log().debug("Generating scenario")
     config = load_yaml(options[CreateOptions.CONFIG])
 
-    cwd = os.getcwd()
-    os.chdir(Path(options[CreateOptions.CONFIG]).parent)
-
     defaults = config["defaults"]
     trace_file = get_trace_file(config, options)
     count = trace_file["total_count"]
     _set_random_seed(defaults, options, trace_file)
     options["scenario_name"] = defaults["base_name"] + f"_{count}"
-    scenario = load_yaml(config["base_template"])
+    scenario = load_yaml(Path(options[CreateOptions.CONFIG].parent, config["base_template"]))
 
     if "create" in config:
         for agent in config["create"]:
-            type_template = load_yaml(Path(agent["type_template"]))
+            type_template = load_yaml(Path(options[CreateOptions.CONFIG].parent, agent["type_template"]))
             agent_type_template = type_template["Agent"]
             contract_type_template = type_template.get("Contracts")
             n_to_create = _get_number_of_agents_to_create(agent["count"], options)
@@ -78,7 +77,8 @@ def generate_scenario(options: dict) -> None:
 
     _resolve_identifiers(scenario, options)
     _resolve_ids(scenario)
-    os.chdir(cwd)
+    _update_series_paths(scenario, options, Path(config["base_template"]))
+
     options["scenario_path"] = Path(options[CreateOptions.DIRECTORY], options["scenario_name"] + ".yaml")
     write_yaml(scenario, options["scenario_path"])
 
@@ -216,7 +216,7 @@ def _digest_choose(input_value: str) -> List[Union[int, float, str]]:
 
 
 def _cast_numeric_strings(values: List[str]) -> List[Union[int, float, str]]:
-    """Returns given List of `values` but numerics are casted in their correct type"""
+    """Returns given List of `values` but numerics are cast in their correct type"""
     casted_values = []
     for value in values:
         try:
@@ -346,3 +346,31 @@ def _resolve_identifiers(input_value: Any, options: dict) -> Any:
                     input_value[key][index] = _get_value_from_field(item, options)
         else:
             input_value[key] = _get_value_from_field(value, options)
+
+
+def _update_series_paths(scenario: dict, options: dict, template_dir: Path) -> None:
+    """
+    Appends relative paths directing to `template_dir` for CSV files defined in `scenario`s `Agents`
+    and (optional) `StringSets`
+    """
+    config_dir = Path(options[CreateOptions.CONFIG]).parent
+    output_dir = Path(options[CreateOptions.DIRECTORY])
+    path_to_append = Path(os.path.relpath(config_dir, start=output_dir), template_dir.parent)
+    for agent in scenario["Agents"]:
+        agent = keys_to_lower(agent)
+        _replace_timeseries_path_in(agent.get("Attributes".lower(), {}), path_to_append)
+
+
+def _replace_timeseries_path_in(attributes: dict, template_path: Path) -> None:
+    """Recursively modify timeseries path in-place in given `attributes` to link to `template_path`"""
+    for attribute_name, attribute_value in attributes.items():
+        if isinstance(attribute_value, str):
+            if attribute_value.lower().endswith(".csv"):
+                attributes[attribute_name] = Path(template_path, attribute_value).__str__()
+        elif isinstance(attribute_value, dict):
+            _replace_timeseries_path_in(attribute_value, template_path)
+        elif isinstance(attribute_value, list):
+            for item in attribute_value:
+                _replace_timeseries_path_in(item, template_path)
+        else:
+            log().debug(DEBUG_NO_PATH_TO_BE_REPLACED_IN.format(attribute_name, attribute_value))