+++ CHANGELOG.md
@@ -5,11 +5,16 @@ SPDX-License-Identifier: CC0-1.0 -->
 # [v1.0.0](https://github.com/FEAT-ML/scengen/releases/tag/v1.0.0) Tba. - Channel 
 ## Changed:
 * **Breaking**: Results are written to subfolders using the scenario name, [#49](https://github.com/FEAT-ML/scengen/issues/49)
+* **Breaking**: Series paths are now written relative to scenario `template` file which facilitates clear separation between static input data and generated scenarios, [#54](https://github.com/FEAT-ML/scengen/issues/54)
 * **Breaking**: Replaced general `range` option with specific `range_int` and `range_float` options, [#56](https://github.com/FEAT-ML/scengen/issues/56)
 * **Breaking**: Renamed section `Agents` to `Agent` in `TypeTemplate` to clarify that one agent type per template is expected, [#58](https://github.com/FEAT-ML/scengen/issues/58)
 * **Breaking**: Renamed replacement identifier from `//THISAGENT` to `//THIS_AGENT` for better readability, [#60](https://github.com/FEAT-ML/scengen/issues/60)
+* **Breaking**: Update minimum version of dependency `amirispy>=2.2,<3`, [#54](https://github.com/FEAT-ML/scengen/issues/54)
 * Update to new logging implementation allowing more flexible and consistent use of logging, [#55](https://github.com/FEAT-ML/scengen/issues/55)
 * `setup.py` to `pyproject.toml`, [#51](https://github.com/FEAT-ML/scengen/issues/51)
+
+## Fixed:
+* Fix `-o/--output-options` documentation on how to use `fameio` output options, [#54](https://github.com/FEAT-ML/scengen/issues/54)
 
 ## Added:
 * New `trace_file` is generated and written to disk if not defined in `GeneratorConfig` or if not found in provided path, [#53](https://github.com/FEAT-ML/scengen/issues/53)